Search sorting by custom column
The sort boolean option for a search column appears to have never been created. It only ever sorts ASC by internalid regardless,This fixes that.
Column can also be set ASC or DESC
Descending: 'sort' = 'true'
Ascending: 'sort' = 'false'
Was not sure if to use string or bool , so went with string to fit in with the rest of the options. so 'true' works but true does not, this could be changed.
If a column is set to sort, then the default internalid sort column is not added.
With this patch multiple columns can be set to sort, though i have no idea if this works or not, compared with the readme stated action (which does not exist) to only sort by last column set to sort.