upsert create item sublist avoid SSS_INVALID_SUBLIST_OPERATION
trying to add an item sublist row to for example a sales order can return a SSS_INVALID_SUBLIST_OPERATION error, this is due to dynamic mode. way around it is to use setCurrentLineItemValue etc instead. this patch just checks for param bySelect in the sublist and uses this method instead, there may be other sublist types that need this, and a more comprehensive integration of these methods, but it works for me for now and thats enough for now...